feat: add metadata for each page
