+++ src/app/layout.tsx
@@ -2,12 +2,12 @@ import "./globals.css";
 import { Inter } from "next/font/google";
 import Footer from "../../components/Footer";
 import Header from "../../components/Header";
+import { Metadata } from "next";
 
 const inter = Inter({ subsets: ["latin"] });
 
-export const metadata = {
-  title: "arts-olympiad",
-  description: "Generated by create next app",
+export const metadata: Metadata = {
+  description: "Schoolchildren everywhere can participate in a global art contest on the theme, My Favorite Sport, and rally the votes to win the gold!"
 };
 
 export default function RootLayout({
@@ -17,6 +17,11 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
+      <link
+        rel="icon"
+        href="/banner-image.svg"
+        type="image/svg+xml"
+      />
       <body className={`${inter.className} bg-neutral-white`}>
         <Header/>
         <main className="flex justify-center">
