+++ src/app/shared-metadata.ts
@@ -1,0 +1,6 @@
+import BannerImage from "../../public/banner-image.png";
+
+export const sharedOpenGraph = {
+  description: "Schoolchildren everywhere can participate in a global art contest on the theme, My Favorite Sport, and rally the votes to win the gold!",
+  images: [ BannerImage.src ]
+};
